refactor(parser): share object data line parsing with control-flow capture
Move the object-line interpretation rule (channel 02 measure length +
non-zero token stream) into event-utils buildBmsObjectLineEntry so the
strict parse path and the #RANDOM / #IF / #SWITCH capture path can no
longer drift apart. Output shapes are unchanged.
